DAOS-4512 rebuild: do not retry for rebuild IV
Let's not retry for rebuild IV and pool prop IV
propagate for the moment. Because once retry,
it may cause rebuild in endless loop, unless
the pool map is properly updated, which should
be explicitly by the caller. I will do that
in another patch.

Fix a typo in rebuild_test yml.

Signed-off-by: Di Wang <di.wang@intel.com>