DAOS-16953 tests: fix btree parameter parsing
In btree.sh the -m and -t parameters were not applied correctly
because of the preceding positional argument containing spaces but
not wrapped in quotes

In btree.c an invalid parameter sequence was not detected
due to of the partially manual parsing process.

btr_test_state introduced to provide global parameters to test
(argc/argv) and avoid static variables for that purpose.

Co-authored-by: Jan Michalski <jan-marian.michalski@hpe.com>
Co-authored-by: Tomasz Gromadzki <tomasz.gromadzki@hpe.com>

Signed-off-by: Tomasz Gromadzki <tomasz.gromadzki@hpe.com>